Made the satellite view the default.
Moved the "Context" menu to the right, renamed it as "Layers
and changed the interaction from a pop-up menu to a fixed one. I've
also adapted its appearance to match the "Sat/Reg" selector above it.

When you select one of those layers, there is a tooltip box
that appears as you hover on the map, but it's behind the info panel
at the left. I've moved the tooltip to the top right, beside the map
zooming controls.

Moved the "number of spills" display to the right, and combined
it with the spill symbols legend.

Labelled the Google Docs id number to find it easily later.